Обновление

+++ Frontend/Script/SaveRepository.cs
@@ -1,47 +1,45 @@
 ﻿using System.IO;
+using System.Text.Json;
+using Frontend.Models;
 
 namespace Frontend.Script;
 
 public class SaveRepository
 {
-    private const string filePath = "token.txt";
-    private const string fileIdPath = "id.txt";
+    public const string filePath = "data.json";
+
+    public static void SaveDataToFile(Token token)
+    {
+        string json = JsonSerializer.Serialize(token);
+        File.WriteAllText(filePath, json);
+    }
+
+    public static Token LoadTokenFromFile()
+    {
+        string json = File.ReadAllText(filePath);
+        Token? token = JsonSerializer.Deserialize<Token>(json);
+
+        if (token == null)
+            throw new NullReferenceException("Токен не найден");
+        
+        return token;
+    }
     
-    public static bool SaveToken(string token)
-    {
-        File.WriteAllText(filePath, token);
-        return true;
-    }
-
-    public static bool SaveId(string id)
-    {
-        File.WriteAllText(fileIdPath, id);
-        return true;
-    }
-
-    public static string ReadToken()
-    {
-        return File.ReadAllText(filePath);
-    }
-
-    public static string ReadId()
-    {
-        return File.ReadAllText(fileIdPath);
-    }
-
-    public static void DeleteSave()
-    {
-        File.Delete(filePath);
-        File.Delete(fileIdPath);
-    }
-
-    public static bool CheckToken()
+    public static bool FileExists()
     {
         return File.Exists(filePath);
     }
-
-    public static bool CheckId()
+    
+    public static void DeleteFile()
     {
-        return File.Exists(fileIdPath);
+        if (File.Exists(filePath))
+        {
+            File.Delete(filePath);
+            Console.WriteLine("Файл успешно удален.");
+        }
+        else
+        {
+            Console.WriteLine("Файл не найден.");
+        }
     }
 }
